pyproject.toml
[tool.poetry]
name = "juji"
version = "0.1.0"
description = ""
authors = ["Eric Laugier <eric.laugier@egen.guru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
cryptography = "^41.0.5"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.22"
mariadb = "^1.1.8"
pydantic-settings = "^2.0.3"
fastapi-mail = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
